Fixing error to consult the loan status before to get the loan requestes

+++ domain/model/src/main/java/co/com/crediya/model/loanrequest/gateways/LoanRequestRepository.java
@@ -6,5 +6,5 @@ import reactor.core.publisher.Mono;
 
 public interface LoanRequestRepository {
     Mono<Void> loanRequest(LoanRequest loanRequest);
-    Flux<LoanRequest> findByStatus(int codeStatus, int page, int offset);
+    Flux<LoanRequest> findByStatus(Long codeStatus, int page, int offset);
 }
